Remove CellsRecord
